pyproject.toml
[project]
name = "jefferson-ai"
version = "0.1.0"
description = "Synthetic voter simulation for election modeling"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "supabase>=2.10.0",
    "prefect>=3.0.0",
    "pydantic>=2.10.0",
    "click>=8.1.0",
    "aiohttp>=3.11.0",
    "beautifulsoup4>=4.12.0",
    "pandas>=2.2.0",
    "zhipuai>=0.1.0",  # GLM
    "google-generativeai>=0.8.0",  # Gemini
    "anthropic>=0.40.0",  # Claude (backup)
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "mypy>=1.0.0",
]
geo = [
    "geopandas>=1.0.0",
    "census>=0.8.0",
]

[project.scripts]
jefferson = "src.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
